Move policy to domain layer
