+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/EDI/Errors/Converters/ChangeMasterData/EffectiveDateIsNotAllowedErrorConverter.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.MeteringPoints.Application.ChangeMasterData;
+using Energinet.DataHub.MeteringPoints.Domain.Policies;
 
 namespace Energinet.DataHub.MeteringPoints.Infrastructure.EDI.Errors.Converters.ChangeMasterData
 {
